[MDEP-204] Add unit test for the ExcludeReactorProjectsArtifactFilter
