+++ src/test/java/org/apache/maven/plugins/dependency/TestSkip.java
@@ -1,11 +1,8 @@
 package org.apache.maven.plugins.dependency;
 
 import org.apache.maven.plugin.Mojo;
-import org.apache.maven.plugin.logging.Log;
 
 import java.io.File;
-import java.io.PrintWriter;
-import java.io.StringWriter;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
@@ -166,159 +163,6 @@ public class TestSkip
         mojo.execute();
 
         assertTrue( log.getContent().contains( "Skipping plugin execution" ) );
-    }
-
-    class CapturingLog
-        implements Log
-    {
-        StringBuilder sb = new StringBuilder();
-
-        /** {@inheritDoc} */
-        public void debug( CharSequence content )
-        {
-            print( "debug", content );
-        }
-
-        /** {@inheritDoc} */
-        public void debug( CharSequence content, Throwable error )
-        {
-            print( "debug", content, error );
-        }
-
-        /** {@inheritDoc} */
-        public void debug( Throwable error )
-        {
-            print( "debug", error );
-        }
-
-        /** {@inheritDoc} */
-        public void info( CharSequence content )
-        {
-            print( "info", content );
-        }
-
-        /** {@inheritDoc} */
-        public void info( CharSequence content, Throwable error )
-        {
-            print( "info", content, error );
-        }
-
-        /** {@inheritDoc} */
-        public void info( Throwable error )
-        {
-            print( "info", error );
-        }
-
-        /** {@inheritDoc} */
-        public void warn( CharSequence content )
-        {
-            print( "warn", content );
-        }
-
-        /** {@inheritDoc} */
-        public void warn( CharSequence content, Throwable error )
-        {
-            print( "warn", content, error );
-        }
-
-        /** {@inheritDoc} */
-        public void warn( Throwable error )
-        {
-            print( "warn", error );
-        }
-
-        /** {@inheritDoc} */
-        public void error( CharSequence content )
-        {
-            System.err.println( "[error] " + content.toString() );
-        }
-
-        /** {@inheritDoc} */
-        public void error( CharSequence content, Throwable error )
-        {
-            StringWriter sWriter = new StringWriter();
-            PrintWriter pWriter = new PrintWriter( sWriter );
-
-            error.printStackTrace( pWriter );
-
-            System.err.println( "[error] " + content.toString() + System.lineSeparator() + System.lineSeparator() + sWriter.toString() );
-        }
-
-        /**
-         * @see org.apache.maven.plugin.logging.Log#error(java.lang.Throwable)
-         */
-        public void error( Throwable error )
-        {
-            StringWriter sWriter = new StringWriter();
-            PrintWriter pWriter = new PrintWriter( sWriter );
-
-            error.printStackTrace( pWriter );
-
-            System.err.println( "[error] " + sWriter.toString() );
-        }
-
-        /**
-         * @see org.apache.maven.plugin.logging.Log#isDebugEnabled()
-         */
-        public boolean isDebugEnabled()
-        {
-            // TODO: Not sure how best to set these for this implementation...
-            return false;
-        }
-
-        /**
-         * @see org.apache.maven.plugin.logging.Log#isInfoEnabled()
-         */
-        public boolean isInfoEnabled()
-        {
-            return true;
-        }
-
-        /**
-         * @see org.apache.maven.plugin.logging.Log#isWarnEnabled()
-         */
-        public boolean isWarnEnabled()
-        {
-            return true;
-        }
-
-        /**
-         * @see org.apache.maven.plugin.logging.Log#isErrorEnabled()
-         */
-        public boolean isErrorEnabled()
-        {
-            return true;
-        }
-
-        private void print( String prefix, CharSequence content )
-        {
-            sb.append( "[" ).append( prefix ).append( "] " ).append( content.toString() ).append( System.lineSeparator() );
-        }
-
-        private void print( String prefix, Throwable error )
-        {
-            StringWriter sWriter = new StringWriter();
-            PrintWriter pWriter = new PrintWriter( sWriter );
-
-            error.printStackTrace( pWriter );
-
-            sb.append( "[" ).append( prefix ).append( "] " ).append( sWriter.toString() ).append( System.lineSeparator() );
-        }
-
-        private void print( String prefix, CharSequence content, Throwable error )
-        {
-            StringWriter sWriter = new StringWriter();
-            PrintWriter pWriter = new PrintWriter( sWriter );
-
-            error.printStackTrace( pWriter );
-
-            sb.append( "[" ).append( prefix ).append( "] " ).append( content.toString() ).append( System.lineSeparator() ).append( System.lineSeparator() ).append( sWriter.toString() ).append( System.lineSeparator() );
-        }
-
-        protected String getContent()
-        {
-            return sb.toString();
-        }
     }
 
 }
